+++ src/main/kotlin/cn/zzwtsy/fc2service/service/Fc2VideoInfoParseService.kt
@@ -1,7 +1,7 @@
 package cn.zzwtsy.fc2service.service
 
-import cn.zzwtsy.fc2service.domain.dto.Fc2VideoInfoDto
+import cn.zzwtsy.fc2service.dto.Fc2VideoInfoDto
 
 interface Fc2VideoInfoParseService {
-    fun parse(fc2Id: Int): Fc2VideoInfoDto?
+    fun parse(fc2Id: Long): Fc2VideoInfoDto?
 }
